Register symbol_to_proc body as a pre-assigned FuncId
Moves the Symbol#to_proc shared body function next to enum_yielder and
yielder in globals/store/function.rs, and registers it up-front in
Globals::new with a compile-time-constant FuncId (FuncId::new(3) via
SYMBOL_TO_PROC_BODY_FUNCID). This removes the OnceLock<FuncId> indirection
previously used by Symbol#to_proc / Proc#call to look up the body's id.

No behavior change; core/symbol spec still reports 5 failures / 0 errors.
