+++ monoruby/src/globals.rs
@@ -211,6 +211,10 @@ impl Globals {
             FuncId::new(2),
             globals.define_builtin_func_rest(OBJECT_CLASS, "", yielder)
         );
+        assert_eq!(
+            SYMBOL_TO_PROC_BODY_FUNCID,
+            globals.define_builtin_func_with(OBJECT_CLASS, "", symbol_to_proc_body, 1, 1, true)
+        );
         globals.random.init_with_seed(None);
         crate::builtins::init_builtins(&mut globals);
         globals
